Template fix for Python 2
This is really really stupid but so it goes.  comment.author will
always be a unicode string, but 'Zack Weinberg' will be an ASCII
string in Python 2.  You can't write u'...' in a Jinja template.
Concatenating comment.author with an empty string causes Jinja to
coerce it to ASCII if it's all-ASCII, making it comparable to the
ASCII string on the RHS.