Handle the contents of the ‘meta’ directory.
This was surprisingly difficult, as there doesn’t seem to be a stock
Metalsmith plugin to rename files based on a frontmatter property,
and the plugin that lets you store frontmatter in sidecar files
(necessary for favicon.ico) eluded several searches.

Also, the ‘pages’ and ‘posts’ top-level directories are now elided
from the output, and we use the default-values plugin to avoid needing
to call layouts twice.  And nodemon notices when index.js changes.