Generate a collection for each topic directory.
This uses metalsmith-auto-collections and is not yet picking up the
metadata from each directory’s index.md as Pelican did.

For this to work correctly, we need to render Markdown *after*
renaming files, and rename_patterns needs to operate on “path”
properties as well as on the actual file names.