Stop using airbnb’s ESLint preferences.
It’s better to manually enable everything *this* project wants above
eslint:recommended.  Tedious to go through the giant list of rules
that aren’t on by default, but will hopefully save annoyances in the
future.

In the process I noticed some places where the code could be cleaned
up:

 - The /u modifier on a regex in JavaScript is more like “use strict”
   than “use unicode”, and should be used basically always
 - If we put eslint in ecmaVersion:2022 mode then we can have a
   top-level “await main()” in index.js instead of the stupid
   .then(() => {}) that technically doesn’t wait for anything
 - A comment in an empty block is enough to make it non-empty
   for purposes of no-empty, use that instead of allowEmptyCatch
 - get rid of a commented-out line that should’ve been deleted ages ago