Accurate date calculation, sorting, and rendering.
Implement a local plugin that pulls commit information out of Git,
filters out the uninteresting changes, and merges that with the
explicit date annotations which may or may not be present on any
given post.  Vivify datestamps using Luxon date objects, making it
possible to reactivate the code in the page templates that
pretty-prints and marks up publication dates.