Add excerpt/teaser/readmore functionality.
This uses a custom clone of metalsmith-more
<https://github.com/kfranqueiro/metalsmith-more/>.  The main reason we
can’t use metalsmith-more itself is, when there is no ‘<!--more-->‘
annotation, metalsmith-more either does nothing or it copies a
_reference_ to the file contents.  In our templates, we want to paste
the entire contents of a file when it doesn’t have a ‘<!--more-->‘
annotation, but, depending on the (unpredictable) order in which the
layout plugin processes files, “the entire contents” might already
include the page chrome!  So it’s essential to take a copy before the
layout plugin starts its work.  I _could_ have written a wrapper but
it seemed like a good opportunity to make this plugin behave more like
the other plugins (in particular, specifying files to operate on with
a multimatch pattern).