Ignore editor scratch and backup files in all subdirectories.
Per the discussion in https://github.com/metalsmith/metalsmith/issues/384
globs passed to metalsmith.ignore are matched using micromatch’s
default settings, which means, among other things, that a pattern with
no slashes in it will match _only_ at the root of the source tree.
(This is intentionally different from how .gitignore works.)  Change
all ignore patterns that were intended to match only the basename
(i.e. no matter where the file is in the tree, it will be ignored)
to begin with ‘**/‘ to get the intended behavior.