Restore the top-level index and archive pages.
The metalsmith-pagination plugin works for us with only a little
additional trickery.  The existing archive template could be reused
mostly intact.  The templates now have access to a filter “canon_path”
that makes sure links from page to page always end with ‘/’ and not
with ‘/index.html’.  (The trailing slash avoids an HTTP redirect in
some cases.)