Fix the last few broken internal links.
404.md was just being sloppy (the link would have worked, but with an HTTP
redirect that can be avoided by writing it differently); domestic-ghosts.md
was just wrong; the-great-2010-ceramic-objects-giveaway.md had page-relative
links, which work on the page itself but not when the excerpt is embedded
in a category or archive page.  I’d _like_ to write a plugin that converts
page-relative to site-relative links but that’s too much work for right now.

Use the metalsmith-broken-link-checker plugin to ensure no future
internal broken links.  (It doesn’t check external links, which means
it’s fast and reliable enough to run all the time in development.)