Switch from markdown-it-katex to markdown-it-dollarmath.
The advantage of -dollarmath is that it *doesn’t* pin our KaTeX
dependency.  We are responsible for providing the math renderer.
This means we can be certain that ‘import katex from "katex"‘ in
our code is the same version of KaTeX that’s actually in use,
which will be important when we implement self-hosting of the
style and font files.